Refactored ExportTrips and Notifications to share one checkin-method
