+++ tests/Feature/NotificationsTest.php
@@ -7,7 +7,6 @@ use App\Models\Like;
 use App\Models\User;
 use Carbon\Carbon;
 use DateTime;
-use Exception;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Notifications\DatabaseNotification;
 use Tests\TestCase;
@@ -22,60 +21,6 @@ class NotificationsTest extends TestCase
         parent::setUp();
 
         $this->user = $this->createGDPRAckedUser();
-    }
-
-    /**
-     * This is mostly copied from Checkin Test and exactly copied from ExportTripsTest.
-     * @param $stationname
-     * @param $ibnr
-     * @param DateTime $now
-     * @param User|null $user
-     * @throws Exception
-     */
-    protected function checkin($stationname, $ibnr, DateTime $now, User $user = null) {
-        if ($user == null) {
-            $user = $this->user;
-        }
-        $trainStationboard = TransportController::TrainStationboard($stationname,
-                                                                    Carbon::createFromTimestamp($now->format('U')),
-                                                                    'express');
-
-        $countDepartures = count($trainStationboard['departures']);
-        if ($countDepartures == 0) {
-            $this->markTestSkipped("Unable to find matching trains. Is it night in $stationname?");
-            return;
-        }
-
-        // Second: We don't like broken or cancelled trains.
-        $i = 0;
-        while ((isset($trainStationboard['departures'][$i]->cancelled)
-                && $trainStationboard['departures'][$i]->cancelled)
-            || count($trainStationboard['departures'][$i]->remarks) != 0) {
-            $i++;
-            if ($i == $countDepartures) {
-                $this->markTestSkipped("Unable to find unbroken train.
-                Is it stormy in $stationname?");
-                return;
-            }
-        }
-        $departure = $trainStationboard['departures'][$i];
-        CheckinTest::isCorrectHafasTrip($departure, $now);
-
-        // Third: Get the trip information
-        $trip = TransportController::TrainTrip(
-            $departure->tripId,
-            $departure->line->name,
-            $departure->stop->location->id
-        );
-
-        // WHEN: User tries to check-in
-        $this->actingAs($user)
-             ->post(route('trains.checkin'), [
-                 'body'        => 'Example Body',
-                 'tripID'      => $departure->tripId,
-                 'start'       => $ibnr,
-                 'destination' => $trip['stopovers'][0]['stop']['location']['id'],
-             ]);
     }
 
     /** @test */
